ingestion-py/requirements.txt
uvicorn==0.34.0
yfinance==1.2.0
fastapi==0.115.6
pydantic==2.12.4
psycopg2-binary==2.9.9